add reset api and update test case
