+++ infrastructure/httpapi/handlers/validators.go
@@ -461,6 +461,47 @@ func (handler *Validators) ListActivities(ctx *fasthttp.RequestCtx) {
 	httpapi.SuccessWithPagination(ctx, blocks, paginationResult)
 }
 
+func (handler *Validators) ResetValidatorStatus(ctx *fasthttp.RequestCtx) {
+	addressParams, addressParamsOk := URLValueGuard(ctx, handler.logger, "address")
+	if !addressParamsOk {
+		return
+	}
+	var identity validator_view.ValidatorIdentity
+	if strings.HasPrefix(addressParams, handler.validatorAddressPrefix) {
+		identity = validator_view.ValidatorIdentity{
+			MaybeOperatorAddress: &addressParams,
+		}
+	} else if strings.HasPrefix(addressParams, handler.consNodeAddressPrefix) {
+		identity = validator_view.ValidatorIdentity{
+			MaybeConsensusNodeAddress: &addressParams,
+		}
+	} else {
+		httpapi.BadRequest(ctx, errors.New("invalid address"))
+		return
+	}
+
+	rawValidator, err := handler.validatorsView.FindBy(identity)
+	if err != nil {
+		if errors.Is(err, rdb.ErrNoRows) {
+			httpapi.NotFound(ctx)
+			return
+		}
+		handler.logger.Errorf("error finding validator by operator address: %v", err)
+		httpapi.InternalServerError(ctx)
+		return
+	}
+
+	rawValidator.Status = constants.BONDED
+
+	if err = handler.validatorsView.Update(rawValidator); err != nil {
+		handler.logger.Errorf("error updating validator by operator address: %v", err)
+		httpapi.InternalServerError(ctx)
+		return
+	}
+
+	httpapi.Success(ctx, rawValidator)
+}
+
 type ValidatorDetails struct {
 	*validator_view.ValidatorRow
